chore(server): simpler parsing of the form to edit a list
Thanks to the use of github.com/go-playground/form, we can parse the
form directly to the target format, without needing the intermediate
one.

I am not very convinced by the use of Alpine.js through. It is only
useful to implement the behavior of the "add element" button, and now it
means we have javascript (alpine.js) templating inside a go template :s